requirements.txt
Brlapi==0.6.1
Django==1.11
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
Flask-Script==2.0.6
Flask-WTF==0.14.2
Jinja2==2.9.6
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
Pillow==2.3.0
PyICU==1.5
PyYAML==3.12
SQLAlchemy==1.1.11
WTForms==2.1
Werkzeug==0.12.2
alembic==0.9.3
apt-clone==0.2.1
beautifulsoup4==4.6.0
bleach==2.0.0
blinker==1.4
chardet==2.2.1
click==6.7
colorama==0.2.5
command-not-found==0.3
defer==1.0.6
diff-match-patch==20121119
dirspec==13.10
django-bootstrap3==9.1.0
django-crispy-forms==1.7.0
django-import-export==0.6.1
dominate==2.3.1
et-xmlfile==1.0.1
gunicorn==19.7.1
html5lib==0.999999999
httplib2==0.8
itsdangerous==0.24
jdcal==1.3
louis==2.5.3
lxml==3.3.3
oauthlib==0.6.1
odfpy==1.3.6
oneconf==0.3.7.14.04.1
openpyxl==2.4.9
pexpect==3.1
piston-mini-client==0.7.5
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
python-apt==0.9.3.5ubuntu2
python-dateutil==2.6.1
python-debian==0.1.21-nmu2ubuntu2
python-dotenv==0.6.5
python-editor==1.0.3
pytz==2017.3
pyxdg==0.25
redis==2.10.6
reportlab==3.0
requests==2.2.1
six==1.10.0
tablib==0.12.1
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.7.1
virtkey==0.63.0
visitor==0.1.3
webencodings==0.5.1
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
